requirements.txt
pyjwt
python-dotenv
streamlit
requests
openai
llama-index
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
nltk
pypdf